server/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
torch
pillow
opencv-python-headless
numpy
python-dotenv
